Receive LTU launch data over FCOM instead of calculating it in the BLD (#3)
Instead of calculating the launch here in the BLD, we now rely on
FastFeedback STATES broadcast. This simplifies BLD a little bit by
removing the calculation from the code. To make this work I added a new
entry to the BLD Pulse Blobs array, a new AVAIL mask, and changed the
fcom stats substitutions file appropriately.

I left the old BPM position entries in the pulse blobs array to avoid
breaking changes to other parts of the multicast code.